Remove custom config from postgresql subchart
The `postgresql` key in `values.yaml` is for overriding the PostgreSQL
chart's default values. It's not a good place to define custom keys
because these are conflated with the actual configuration from the
subchart. This changes removes those custom configs.

It also removes any configured values that already matched the default
from the PostgreSQL subchart, and were therefore not doing anything.