feat(prefect-server): support sqlite database (#418)
* feat(prefect-server): support sqlite database

Supports using sqlite for the database in place of postgresql.

Closes https://linear.app/prefect/issue/PLA-139/support-sqlite-database

* Disable hpa, secret if sqlite enabled

- HPA is not desired given that sqlite requires 1 replica
- secret is postgresql-specific

* Add tests for sqlite settings

* Fix resiliency of separate test

This test, separate from my changes, failed because it was looking for a
volumeMount at a specific index. This change allows it to just confirm
the entry is present without worrying about a specific location in the
list.

* Fix mount setting, hosting at /data

The application was failing to start, saying it couldn't access the
database file.

This change mounts a directory without a subPath setting, allowing the
application to create the file instead.

* Make persistence settings more configurable

* Add to values schema

* Add documentation on enabling sqlite

* Replace hard-coded PVC name

Uses a template to calculate the PVC name, matching the pattern used for
other objects like the Deployment.