chore(docs): document how to set imagePullSecrets in baseJobTemplate (#446)
* Document how to set imagePullSecrets in baseJobTemplate

Documents how to set imagePullSecrets in the baseJobTemplate.

Will be expanded as part of https://linear.app/prefect/issue/PLA-1020/expand-documentation-on-how-to-use-basejobtemplate

Closes https://linear.app/prefect/issue/PLA-1029/document-how-to-set-image-pull-secrets-in-the-base-job-template

* Remove '.Values' from docs

'.Values' is how values are referenced from within Helm templates, not
how end users set it.

* Fix JSON formatting of imagePullSecrets

* Clarify that configuration is not merged with default

* Add back description of base job template

* Remove warning about BJT updates being ignored

Now that we have the initContainer that calls `prefect work-pool
update`, updates are indeed respected.