fix(ci): resolve chart dependencies before unittest
The helm-unittest action does its best to resolve dependencies for each
chart being tested, but it seems this doesn't always work.

We noticed some unit test failures claiming to not have access to a
template that comes from the PostgreSQL chart.

To address that, we'll manually resolve the dependencies beforehand to
be sure they're all resolved.

Closes https://linear.app/prefect/issue/PLA-1222