fix(ci): pin `helm-unittest` version (#469)
* fix(ci): resolve chart dependencies before unittest

The helm-unittest action does its best to resolve dependencies for each
chart being tested, but it seems this doesn't always work.

We noticed some unit test failures claiming to not have access to a
template that comes from the PostgreSQL chart.

To address that, we'll manually resolve the dependencies beforehand to
be sure they're all resolved.

Closes https://linear.app/prefect/issue/PLA-1222

* Align helm version to 3.17

* Pin to helm-unittest 0.7.2

Version 0.8 includes a bug that led to this error:
https://github.com/helm-unittest/helm-unittest/issues/588

For now, we'll pin to 0.7.2, the last known-good version.

* Revert 'make buildall' step

No longer needed given the helm-unittest version pin.