feat: add value for deployment.spec.strategy
Adds value `server.updateStrategy`.

Useful when using sqlite database backend on a mounted volume, where the
volume can only be mounted on a single node at a time. If the strategy
type is `RollingUpdate` and the new pod is scheduled on a different node,
the deployment will hang indefinitely, as the new pod is not able to
start as long as the disk it requires is mounted to the old one.