feat(prefect-server): make Deployment strategy configurable (#493)
* feat: add value for deployment.spec.strategy.type

Adds value `server.deploymentStrategyType`.

Useful when using sqlite database backend on a mounted volume, where the
volume can only be mounted on a single node at a time. If the strategy
is "OrderedReady" and the new pod is scheduled on a different node, the
deployment will hang indefinitely.

* feat: add value for deployment.spec.strategy

Adds value `server.updateStrategy`.

Useful when using sqlite database backend on a mounted volume, where the
volume can only be mounted on a single node at a time. If the strategy
type is `RollingUpdate` and the new pod is scheduled on a different node,
the deployment will hang indefinitely, as the new pod is not able to
start as long as the disk it requires is mounted to the old one.

* update values schema with new additions

* use string values in examples to match schema

* fix wrong variable value reference

* add test for .spec.strategy.rollingUpdate

---------

Co-authored-by: Mitchell Nielsen <mitchnielsen@users.noreply.github.com>